spike: rename imports
