+++ packages/plugin-core/src/components/lookup/LookupProviderV3.ts
@@ -21,7 +21,8 @@ import { LookupControllerV3 } from "./LookupControllerV3";
 import { NoteLookupCommand } from "../../commands/NoteLookupCommand";
 import { DendronQuickPickerV2, DendronQuickPickState } from "./types";
 import { OldNewLocation, PickerUtilsV2 } from "./utils";
-import { NotePickerUtils, SchemaPickerUtils } from "../lookup/NotePickerUtils";
+import { NotePickerUtils } from "../lookup/NotePickerUtils";
+import { SchemaPickerUtils } from "../lookup/SchemaPickerUtils";
 import { transformQueryString } from "./queryStringTransformer";
 import stringSimilarity from "string-similarity";
 import { IDendronQuickInputButton } from "./ButtonTypes";
